requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
PyJWT==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-env==1.0.0
pytz==2022.4
requests==2.28.1
six==1.16.0
twilio==7.14.2
urllib3==1.26.12
